backend/requirements.txt
# Alinhave Backend - dependências
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.0
pydantic>=2.0.0
python-multipart>=0.0.6
docxtpl>=0.16.0
python-docx>=1.0.0
mammoth>=1.6.0
htmldocx>=0.0.6
beautifulsoup4>=4.7.0
